Fixed darkmode and added pleasant styling. The locked in filter should be recognizable from the one that is currently being written. Therefore I used the powerful blue to emphasize that the locked in filter is in fact in use. Also changed the shade of gray because the contrast wasn't accessible. Made the background white (I think it looks more pleasing).
I added some horizontal padding on the bubbles to make the negative space more balanced + added a right margin to create a small gap between each filter for a visually pleasing look.

I also made the font-size the same as in the search text, so that there is no jump between font-sizes when the text is recognized as a filter. Feature is self explanatory. I think this is smoother